+++ src/main/java/racingcar/domain/CarRacingManager.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.view.InputView;
