+++ src/main/java/racingcar/domain/GoForwardOrStop.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class GoForwardOrStop {
     private static final int MIN_GO_FORWARD_VALUE = 3;
