+++ src/main/java/racingcar/domain/RacingTimeRegister.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.exception.InvalidRacingTimeException;
