+++ src/test/java/racingcar/CarRepositoryTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.CarRepository;
 
 import java.util.ArrayList;
 import java.util.List;
