+++ src/test/java/racingcar/RacingTimeRegisterTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RacingTimeRegister;
 import racingcar.exception.InvalidRacingTimeException;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
