+++ src/test/java/racingcar/RandomGeneratorTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.RandomGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
